pyproject.toml
[tool.poetry]
name = "inventox"
version = "0.1.0"
description = ""
authors = ["Dennis Gera <geradennis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
sqlalchemy = "^2.0.35"
uvicorn = "^0.30.6"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
alembic = "^1.13.3"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
